Renamed things to reflect what they now do, and updated import statements.
