+++ lib/improver/cli/extend_radar_mask.py
@@ -32,7 +32,7 @@
 """Script to extend a radar mask based on coverage data."""
 
 from improver.argparser import ArgParser
-from improver.metadata.enforce_datatypes_units import check_cube_not_float64
+from improver.metadata.check_datatypes import check_cube_not_float64
 from improver.nowcasting.utilities import ExtendRadarMask
 from improver.utilities.load import load_cube
 from improver.utilities.save import save_netcdf
