+++ lib/improver/tests/metadata/test_check_datatypes.py
@@ -28,7 +28,7 @@
 # CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
-"""Unit tests for the improver.metadata.enforce_datatypes_units module."""
+"""Unit tests for the improver.metadata.check_datatypes module."""
 
 import unittest
 from datetime import datetime
@@ -38,7 +38,7 @@ import numpy as np
 from iris.coords import AuxCoord
 from iris.tests import IrisTest
 
-import improver.metadata.enforce_datatypes_units as enforce
+import improver.metadata.check_datatypes as enforce
 from improver.tests.set_up_test_cubes import (
     set_up_variable_cube, set_up_probability_cube, set_up_percentile_cube)
